add logic middleware channel connection
